fix add meaure to historyRepositoryImpl
