+++ .gitignore
@@ -73,6 +73,7 @@ ensime.sbt
 **/**/.bloop
 **/**/.metals
 **/**/.tag?
+**/**/.TAG?
 TAGS
 node-client/__pycache__
 node-client/ve
